[refactor] tea commands to be in their own module.
